Randomly space out projectiles
Previously they were right close to each other, now each one has a
spacing amount for how far the next one should be.

+++ game/game.go
@@ -56,7 +56,7 @@ type GameScreen struct {
 	Box         *Box
 	Dusts       Dusts
 	Projectiles Projectiles
-	Tick        int64
+	Tick        int
 	TouchIDs    *[]ebiten.TouchID
 }
